+++ libs/@rucken/core/src/shared/models/group.ts
@@ -28,12 +28,16 @@ export class Group implements IModel {
   @Type(serializeModel(Permission))
   permissions: Permission[] = [];
   get permissionsAsString() {
-    const permissionsLength = this.permissions.length;
-    if (permissionsLength > 14) {
-      return this.permissions.filter((item, index) => index < 7).join(', ') +
-        ((permissionsLength > 7) ? `... +${permissionsLength - 7}` : '');
+    if (this.permissions) {
+      const permissionsLength = this.permissions.length;
+      if (permissionsLength > 14) {
+        return this.permissions.filter((item, index) => index < 7).join(', ') +
+          ((permissionsLength > 7) ? `... +${permissionsLength - 7}` : '');
+      }
+      return this.permissions.join(', ');
+    } else {
+      return '';
     }
-    return this.permissions.join(', ');
   }
   toString() {
     return this.title;