+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+<a name="2.1.0"></a>
+# [2.1.0](https://github.com/rucken/core/compare/2.0.9...2.1.0) (2018-05-08)
+
+
+### Bug Fixes
+
+* **perm.pipe:** Add arguments for pipe: defaultValue: boolean = true, trueValue: any = true, falseValue: any = false ([4189e83](https://github.com/rucken/core/commit/4189e83))
+* Add check variables to undefined ([6770db2](https://github.com/rucken/core/commit/6770db2))
+
+
+### Features
+
+* **pipes:** Add userPerm pipe, for check access to object ([92f06e4](https://github.com/rucken/core/commit/92f06e4))
+
+
+
 <a name="2.0.9"></a>
 ## [2.0.9](https://github.com/rucken/core/compare/2.0.8...2.0.9) (2018-04-29)